Drop model selection from subscription auth; STRIX_LLM stays the single source of truth
- remove strix models / auth model and the openai/subscription pseudo-model
- remove the hardcoded model catalog, default model, and live catalog fetch
- subscription routing: signed in + openai/<model> + no LLM_API_KEY
- merge strix/auth into strix/config/codex.py; auth CLI is login/status/logout only
- clearer guardrail/backend error copy
