+++ tests/test_auth_cli.py
@@ -2,12 +2,11 @@
 
 from __future__ import annotations
 
-import os
 from typing import TYPE_CHECKING, Any
 
 import pytest
 
-from strix.auth import codex, store
+from strix.config import codex
 from strix.interface import auth_cli
 
 
@@ -17,7 +16,7 @@ if TYPE_CHECKING:
 
 @pytest.fixture(autouse=True)
 def _tmp_store(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(store, "AUTH_PATH", tmp_path / "home" / ".strix" / "subscription-auth.json")
+    monkeypatch.setattr(codex, "AUTH_PATH", tmp_path / "home" / ".strix" / "subscription-auth.json")
 
 
 def test_login_provider_is_chatgpt() -> None:
@@ -81,26 +80,8 @@ def test_finish_rejects_missing_code() -> None:
     assert exc.value.code == "no_code"
 
 
-def test_select_model_requires_arg() -> None:
-    assert auth_cli.run_auth(["model"]) == 2
-
-
-def test_select_model_valid_persists_env(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(codex, "is_authenticated", lambda: True)
-    monkeypatch.setattr(codex, "refresh_subscription_models", lambda: codex.SUBSCRIPTION_MODELS)
-    monkeypatch.setattr(auth_cli, "persist_current", lambda: None)
-    monkeypatch.setenv("STRIX_LLM", "openai/subscription")
-
-    assert auth_cli.run_auth(["model", "gpt-5.5"]) == 0
-    assert os.environ["STRIX_LLM"] == "openai/subscription/gpt-5.5"
-
-
-def test_select_model_rejects_unknown(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(codex, "is_authenticated", lambda: True)
-    monkeypatch.setattr(codex, "refresh_subscription_models", lambda: codex.SUBSCRIPTION_MODELS)
-    monkeypatch.setattr(auth_cli, "persist_current", lambda: None)
-
-    assert auth_cli.run_auth(["model", "gpt-4o"]) == 2
+def test_model_subcommand_removed() -> None:
+    assert auth_cli.run_auth(["model", "gpt-5.5"]) == 2
 
 
 @pytest.mark.parametrize("provider", ["chatgpt", "codex", "ChatGPT"])
@@ -120,7 +101,6 @@ def test_login_accepts_provider_aliases(provider: str, monkeypatch: pytest.Monke
 
     monkeypatch.setattr(auth_cli, "_run_oauth_flow", _fake_flow)
     monkeypatch.setattr(codex, "save_record", lambda _record: None)
-    monkeypatch.setattr(auth_cli, "_persist_subscription_config", lambda: None)
 
     assert auth_cli.run_auth(["login", provider]) == 0
     assert reached["flow"] is True
