+++ tests/test_model_retry.py
@@ -13,7 +13,7 @@ import asyncio
 
 from agents.retry import ModelRetryNormalizedError, RetryPolicyContext
 
-from strix.auth import codex
+from strix.config import codex
 from strix.config.models import DEFAULT_MODEL_RETRY, _retry_statusless_provider_errors
